requirements.txt
requests
mysql-connector-python
Flask  # Only if using Flask for the API
python-dotenv